requirements.txt
pandas
numpy
pytz
scipy
cognite-logger
python-dateutil
pydantic